Replace Into with From for built-in types.
https://doc.rust-lang.org/std/convert/trait.Into.html

Prior to Rust 1.41, if the destination type was not part of the current
crate then you couldn't implement From directly. This will fail to
compile in older versions of the language because Rust's orphaning rules
used to be a little bit more strict. To bypass this, you could implement
Into directly.

It is important to understand that Into does not provide a From
implementation (as From does with Into). Therefore, you should always
try to implement From and then fall back to Into if From can't be
implemented.

Signed-off-by: Sean Wilson <spwilson27@gmail.com>