feat: update to 2021
This commit introduces the edition 2021 for the crate, and perform some
tweaks.

Prior to this commit, the crate was `std`, and didn't implement `Deref`
for the wrapped types. Implementing `Deref` is often beneficial -
especially for base types such as integers. It will allow the user to
have the native functionality of the integers while operating with the
wrapped type.

The wrapped operations will naturally result in native integers and they
will have to be reconstructed to the target endianness before they are
returned.

Also, the crate was requiring `std`, when this is not actually needed.
We perform just basic operations with memory, and these are fully
supported by `core`. It will allow the types to be represented and
operated under constrained environments such as WASM.